+++ src/Messaging.Shared/Protocol/ProtocolBase.cs
@@ -7,11 +7,11 @@ public abstract class ProtocolBase : IMessageProtocol {
 
     public abstract Task<bool> ProcessAsync(MessageData message);
 
-    protected async Task EnqueueAck(MessageConnectionHandler handler, StringIdentifier source, StringIdentifier target, int idToAck) {
+    protected async Task EnqueueAck(MessageConnectionHandler handler, StringIdentifier source, StringIdentifier target, long idToAck) {
         await handler.WriteToOutBufferAsync(CreateAck(source, target, idToAck));
     }
 
-    public virtual MessageData CreateAck(StringIdentifier source, StringIdentifier target, int idToAck) {
+    public virtual MessageData CreateAck(StringIdentifier source, StringIdentifier target, long idToAck) {
         MessageData message = new() {
             Id = idToAck,
             Type = MessageType.Ack,
